requirements.txt
# ── Core LLM + RAG ──────────────────────────────────────────────────────────
vanna>=0.7.0
chromadb>=0.4.0
ollama>=0.3.0

# ── Frontend ─────────────────────────────────────────────────────────────────
streamlit>=1.35.0

# ── Database ─────────────────────────────────────────────────────────────────
pyodbc>=5.1.0
sqlalchemy>=2.0.0

# ── Data processing ─────────────────────────────────────────────────────────
pandas>=2.0.0
numpy>=1.24.0

# ── Chart export (Vega-Lite → PNG/SVG) ──────────────────────────────────────
vl-convert-python>=1.1.0

# ── Configuration ───────────────────────────────────────────────────────────
pyyaml>=6.0
python-dotenv>=1.0.0